requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
black==20.8b1
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
distlib==0.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
identify==1.5.10
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.5.0
pathspec==0.8.1
pre-commit==2.9.3
psycopg2==2.8.6
PyJWT==2.0.0
pylint==2.6.0
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
wrapt==1.12.1
